Move Kafka to Producer module
